Remove post from queue when not indexable
This change fixes updates that trigger multiple filters for the same
post. Currently, when any of the filters adds the post to the update
queue the post is indexed, even when other filters determine
that the post should not be indexed. This change makes the last
filter have the final word about this.